Fix a masked binding bug in AffLin?
tests/good/discard_prods.ppl was failing in my WIP because AffLin turned
    `let p=(id,id) in y`
into
    `let p=(id,id) in let (x,x0)=p in let <_,x>=x0 in let <_,x>=x in y`
which is wrong because `let <_,x>=x0` should be `let <_,x1>=x0` or `let <_,x0>=x0`.
Both of the TmElimAdditive binders should be generated with `x` and `x0` in the Ctxt.
I guess this problem was masked because old Subst code renamed more often.
Anyway, it turns out that order of operations matters.